Issue 69.1: extract daemon-loop helpers to server/helpers.rs
main.rs goes from 845 → 317 prod LOC by relocating 20 daemon-loop
helper fns (~530 LOC) into a new sibling module
userspace-dp/src/server/helpers.rs.

Helpers moved (called by both `run()` and the handle_stream handlers):

Status / sync:
- refresh_status (77 LOC)
- forwarding_unsupported_error (10)
- build_synced_session_key (18)
- build_synced_session_entry (111)
- parse_session_sync_mac (18)

Binding orchestration:
- reconcile_status_bindings (22)
- should_run_afxdp (3)
- set_bindings_forwarding_armed (6)
- wait_for_binding_settle (10)
- bindings_settled (8)
- same_binding_plan (4)
- snapshot_binding_plan_key (36)
- include_userspace_binding_interface (17)
- replan_queues (51)
- replan_bindings_from_candidates (50)
- summarize_queues (28)
- is_userspace_candidate_interface (3)
- linux_ifname (3)
- rx_queue_count (12)

Persistence:
- write_state (22)

All widened from file-private `fn` to `pub(crate) fn` so main.rs's
`use server::helpers::*;` and handlers.rs's `use super::super::*;`
chains continue to resolve.

server/mod.rs adds:
    pub(crate) mod helpers;

(no glob re-export at server::* — main.rs reaches helpers directly
via `server::helpers::*` and handlers.rs reaches them via
`super::super::*` through the crate root.)

main.rs now contains only:
- `fn main()` and `fn run()` (the daemon entry point + main loop)
- `fn parse_args()`
- `fn derive_session_socket_path` / `fn derive_event_socket_path`
- module declarations + use statements

Pure relocation. Bodies byte-for-byte identical.

cargo test --release: 865 passed, 0 failed (warning count back to 92
baseline after dropping the unused `pub(crate) use helpers::*;` glob
re-export from server/mod.rs).
